+++ packages/expo-modules-core/src/NativeViewManagerAdapter.native.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
-import { NativeModules, requireNativeComponent, HostComponent } from 'react-native';
+import { findNodeHandle, NativeModules, requireNativeComponent, HostComponent } from 'react-native';
+
+import { requireNativeModule } from './requireNativeModule';
 
 // To make the transition from React Native's `requireNativeComponent` to Expo's
 // `requireNativeViewManager` as easy as possible, `requireNativeViewManager` is a drop-in
@@ -56,14 +58,40 @@ export function requireNativeViewManager<P>(viewName: string): React.ComponentTy
     requireCachedNativeComponent<NativeExpoComponentProps>(reactNativeViewName);
   const proxiedPropsNames = viewManagerConfig?.propsNames ?? [];
 
-  // Define a component for universal-module authors to access their native view manager
-  const NativeComponentAdapter = React.forwardRef<any>((props, ref) => {
-    const nativeProps = omit(props, proxiedPropsNames);
-    const proxiedProps = pick(props, proxiedPropsNames);
-    return <ReactNativeComponent {...nativeProps} proxiedProperties={proxiedProps} ref={ref} />;
-  }) as React.ComponentType<P>;
-  NativeComponentAdapter.displayName = `Adapter<${viewName}>`;
-  return NativeComponentAdapter;
+  class NativeComponent extends React.PureComponent<P> {
+    displayName = viewName;
+
+    // This will be accessed from native when the prototype functions are called,
+    // in order to find the associated native view.
+    nativeTag: number | null = null;
+
+    componentDidMount(): void {
+      this.nativeTag = findNodeHandle(this);
+    }
+
+    render(): React.ReactNode {
+      const nativeProps = omit(this.props, proxiedPropsNames);
+      const proxiedProps = pick(this.props, proxiedPropsNames);
+
+      return <ReactNativeComponent {...nativeProps} proxiedProperties={proxiedProps} />;
+    }
+  }
+
+  const nativeModule = requireNativeModule(viewName);
+  const nativeViewPrototype = nativeModule.ViewPrototype;
+
+  if (nativeViewPrototype) {
+    // Temporarily each function is wrapped to pass the native tag as the argument.
+    // In the future, native will automatically get the tag from `this` and then we can simply use
+    // `Object.assign(NativeComponent.prototype, nativeViewPrototype)` instead.
+    for (const key of Object.getOwnPropertyNames(nativeViewPrototype)) {
+      NativeComponent.prototype[key] = function (...args) {
+        return nativeViewPrototype[key].call(this, this.nativeTag, ...args);
+      };
+    }
+  }
+
+  return NativeComponent;
 }
 
 function omit(props: Record<string, any>, propNames: string[]) {